+++ app/build.gradle
@@ -23,11 +23,6 @@ android {
         }
     }
 
-    compileOptions {
-//        sourceCompatibility JavaVersion.VERSION_11
-//        targetCompatibility JavaVersion.VERSION_11
-    }
-
     kotlin {
         jvmToolchain(11)
     }
