+++ decoder-glide/build.gradle
@@ -15,11 +15,6 @@ android {
         targetSdk rootProject.targetSdkVersion
     }
 
-//    compileOptions {
-//        sourceCompatibility 1.8
-//        targetCompatibility 1.8
-//    }
-
     kotlin {
         jvmToolchain(11)
     }
@@ -30,7 +25,7 @@ android {
 }
 
 dependencies {
-    implementation project(":heifconverter")
+    api project(":decoder")
 
     // Coroutines
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:$coroutines_version"
