+++ heifconverter/build.gradle
@@ -21,11 +21,6 @@ android {
         }
     }
 
-//    compileOptions {
-//        sourceCompatibility 1.8
-//        targetCompatibility 1.8
-//    }
-
     kotlin {
         jvmToolchain(11)
     }
@@ -38,6 +33,9 @@ android {
 dependencies {
     // local
     implementation fileTree(dir: 'libs', include: ['*.jar'])
+
+    implementation project(":decoder")
+    implementation project(":decoder-glide")
 
     // AndroidX
     implementation 'androidx.appcompat:appcompat:1.6.1'
