+++ settings.gradle
@@ -2,6 +2,7 @@ rootProject.name = 'HEIFConverter'
 
 include ':heifconverter'
 include ':heifconverter-dsl'
+include ':decoder'
 include ':decoder-glide'
 include ':urlloader-okhttp3'
 include ':app'
